Rename ServiceHandle module.
