+++ overwatch/src/services/runner/service_runner.rs
@@ -7,10 +7,10 @@ use tracing::{error, info};
 use crate::{
     overwatch::handle::OverwatchHandle,
     services::{
-        handle::ServiceHandle,
         lifecycle::LifecycleMessage,
         resources::ServiceResources,
         runner::ServiceRunnerHandle,
+        service_handle::ServiceHandle,
         state::{ServiceState, StateOperator},
         ServiceCore,
     },
